Imported Upstream version 2.2.0~pre1

+++ modules/codec/cdg.c
@@ -2,7 +2,7 @@
  * cdg.c: CDG decoder module
  *****************************************************************************
  * Copyright (C) 2007 Laurent Aimar
- * $Id: 3d132f9700217ce27735c1d63216de798a7730fc $
+ * $Id: 5a28a33799d1d96114577664f90372eb5baf98c0 $
  *
  * Authors: Laurent Aimar <fenrir # via.ecp.fr>
  *
@@ -329,8 +329,8 @@ static int DecodeScroll( decoder_sys_t *p_cdg, const uint8_t *p_data, int b_copy
             }
             else
             {
-                if( dy < 0 || dy >= CDG_SCREEN_HEIGHT ||
-                    dx < 0 || dx >= CDG_SCREEN_WIDTH )
+                if( dy < 0 || (unsigned)dy >= CDG_SCREEN_HEIGHT ||
+                    dx < 0 || (unsigned)dx >= CDG_SCREEN_WIDTH )
                     continue;
             }
             p_cdg->screen[dy*CDG_SCREEN_PITCH+dx] = copy[y*CDG_SCREEN_PITCH+x];
